refactor: extract dictionary experience feature shell
