+++ website/src/features/dictionary-experience/index.js
@@ -1,0 +1,2 @@
+export { default as DictionaryExperience } from "./DictionaryExperience";
+export { useDictionaryExperience } from "./hooks/useDictionaryExperience";
